task size to integer, started validation schema

+++ server/models/Task.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
         },
         swimlaneOrderNumber: DataTypes.INTEGER,
         color: DataTypes.STRING,
-        size: DataTypes.DOUBLE,
+        size: DataTypes.INTEGER,
         difficulty: DataTypes.INTEGER,
         deletedAt: DataTypes.DATE,
     })
